requirements.txt
HEAD
# NomadPuddy

## System Dependencies
Before running the scripts, ensure you have the following system dependencies installed:

- ffmpeg

On Ubuntu or Debian-based systems (including Raspberry Pi OS), you can install it with:

certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.1
openai-whisper==20231117
packaging==24.1
paho-mqtt==2.1.0
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.27.3
pycparser==2.22
regex==2024.7.24
requests==2.32.3
scipy==1.14.0
six==1.16.0
sounddevice==0.5.0
sympy==1.13.2
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.5
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
whisper==1.1.10
4365661 (Initial commit on Radxa with updated MQTT address and tested STT functionality)
